added cmake
