+++ example/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_executable(example main.cpp)
+target_link_libraries(example process)
